fix: banner breaking certain parts of the website
