+++ src/app.scss
@@ -34,7 +34,7 @@ body {
 .wrapper {
 	margin-inline: auto;
 	width: min(90%, 80rem);
-	margin-top: 7rem;
+	margin-top: 1rem;
 }
 
 :root {
