+++ src/routes/patches/+page.svelte
@@ -207,7 +207,6 @@
 	}
 
 	.search {
-		padding-top: 5rem;
 		padding-bottom: 1.25rem;
 		background-color: var(--surface-eight);
 	}
@@ -256,10 +255,6 @@
 
 		aside {
 			display: none;
-		}
-
-		.search {
-			padding-top: 4.5rem;
 		}
 
 		.patches-container {
